@@ -1,3 +1,3 @@
 module fake_nodejs_module // Exclude this directory from Go tools
 
-go 1.16
+go 1.17
