@@ -19,32 +19,35 @@ import (
 // package main
 //
 // import (
-// 	"github.com/pulumi/pulumi-tailscale/sdk/go/tailscale"
-// 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+//
+//	"github.com/pulumi/pulumi-tailscale/sdk/go/tailscale"
+//	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+//
 // )
 //
-// func main() {
-// 	pulumi.Run(func(ctx *pulumi.Context) error {
-// 		sampleDevice, err := tailscale.GetDevice(ctx, &GetDeviceArgs{
-// 			Name: "device.example.com",
-// 		}, nil)
-// 		if err != nil {
-// 			return err
-// 		}
-// 		_, err = tailscale.NewDeviceSubnetRoutes(ctx, "sampleRoutes", &tailscale.DeviceSubnetRoutesArgs{
-// 			DeviceId: pulumi.String(sampleDevice.Id),
-// 			Routes: pulumi.StringArray{
-// 				pulumi.String("10.0.1.0/24"),
-// 				pulumi.String("1.2.0.0/16"),
-// 				pulumi.String("2.0.0.0/24"),
-// 			},
-// 		})
-// 		if err != nil {
-// 			return err
-// 		}
-// 		return nil
-// 	})
-// }
+//	func main() {
+//		pulumi.Run(func(ctx *pulumi.Context) error {
+//			sampleDevice, err := tailscale.GetDevice(ctx, &GetDeviceArgs{
+//				Name: "device.example.com",
+//			}, nil)
+//			if err != nil {
+//				return err
+//			}
+//			_, err = tailscale.NewDeviceSubnetRoutes(ctx, "sampleRoutes", &tailscale.DeviceSubnetRoutesArgs{
+//				DeviceId: pulumi.String(sampleDevice.Id),
+//				Routes: pulumi.StringArray{
+//					pulumi.String("10.0.1.0/24"),
+//					pulumi.String("1.2.0.0/16"),
+//					pulumi.String("2.0.0.0/24"),
+//				},
+//			})
+//			if err != nil {
+//				return err
+//			}
+//			return nil
+//		})
+//	}
+//
 // ```
 type DeviceSubnetRoutes struct {
 	pulumi.CustomResourceState
@@ -148,7 +151,7 @@ func (i *DeviceSubnetRoutes) ToDeviceSubnetRoutesOutputWithContext(ctx context.C
 // DeviceSubnetRoutesArrayInput is an input type that accepts DeviceSubnetRoutesArray and DeviceSubnetRoutesArrayOutput values.
 // You can construct a concrete instance of `DeviceSubnetRoutesArrayInput` via:
 //
-//          DeviceSubnetRoutesArray{ DeviceSubnetRoutesArgs{...} }
+//	DeviceSubnetRoutesArray{ DeviceSubnetRoutesArgs{...} }
 type DeviceSubnetRoutesArrayInput interface {
 	pulumi.Input
 
@@ -173,7 +176,7 @@ func (i DeviceSubnetRoutesArray) ToDeviceSubnetRoutesArrayOutputWithContext(ctx 
 // DeviceSubnetRoutesMapInput is an input type that accepts DeviceSubnetRoutesMap and DeviceSubnetRoutesMapOutput values.
 // You can construct a concrete instance of `DeviceSubnetRoutesMapInput` via:
 //
-//          DeviceSubnetRoutesMap{ "key": DeviceSubnetRoutesArgs{...} }
+//	DeviceSubnetRoutesMap{ "key": DeviceSubnetRoutesArgs{...} }
 type DeviceSubnetRoutesMapInput interface {
 	pulumi.Input
 
@@ -207,6 +210,16 @@ func (o DeviceSubnetRoutesOutput) ToDeviceSubnetRoutesOutput() DeviceSubnetRoute
 
 func (o DeviceSubnetRoutesOutput) ToDeviceSubnetRoutesOutputWithContext(ctx context.Context) DeviceSubnetRoutesOutput {
 	return o
+}
+
+// The device to set subnet routes for
+func (o DeviceSubnetRoutesOutput) DeviceId() pulumi.StringOutput {
+	return o.ApplyT(func(v *DeviceSubnetRoutes) pulumi.StringOutput { return v.DeviceId }).(pulumi.StringOutput)
+}
+
+// The subnet routes that are enabled to be routed by a device
+func (o DeviceSubnetRoutesOutput) Routes() pulumi.StringArrayOutput {
+	return o.ApplyT(func(v *DeviceSubnetRoutes) pulumi.StringArrayOutput { return v.Routes }).(pulumi.StringArrayOutput)
 }
 
 type DeviceSubnetRoutesArrayOutput struct{ *pulumi.OutputState }
