@@ -19,23 +19,26 @@ import (
 // package main
 //
 // import (
-// 	"github.com/pulumi/pulumi-tailscale/sdk/go/tailscale"
-// 	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+//
+//	"github.com/pulumi/pulumi-tailscale/sdk/go/tailscale"
+//	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
+//
 // )
 //
-// func main() {
-// 	pulumi.Run(func(ctx *pulumi.Context) error {
-// 		_, err := tailscale.NewDnsSearchPaths(ctx, "sampleSearchPaths", &tailscale.DnsSearchPathsArgs{
-// 			SearchPaths: pulumi.StringArray{
-// 				pulumi.String("example.com"),
-// 			},
-// 		})
-// 		if err != nil {
-// 			return err
-// 		}
-// 		return nil
-// 	})
-// }
+//	func main() {
+//		pulumi.Run(func(ctx *pulumi.Context) error {
+//			_, err := tailscale.NewDnsSearchPaths(ctx, "sampleSearchPaths", &tailscale.DnsSearchPathsArgs{
+//				SearchPaths: pulumi.StringArray{
+//					pulumi.String("example.com"),
+//				},
+//			})
+//			if err != nil {
+//				return err
+//			}
+//			return nil
+//		})
+//	}
+//
 // ```
 type DnsSearchPaths struct {
 	pulumi.CustomResourceState
@@ -126,7 +129,7 @@ func (i *DnsSearchPaths) ToDnsSearchPathsOutputWithContext(ctx context.Context) 
 // DnsSearchPathsArrayInput is an input type that accepts DnsSearchPathsArray and DnsSearchPathsArrayOutput values.
 // You can construct a concrete instance of `DnsSearchPathsArrayInput` via:
 //
-//          DnsSearchPathsArray{ DnsSearchPathsArgs{...} }
+//	DnsSearchPathsArray{ DnsSearchPathsArgs{...} }
 type DnsSearchPathsArrayInput interface {
 	pulumi.Input
 
@@ -151,7 +154,7 @@ func (i DnsSearchPathsArray) ToDnsSearchPathsArrayOutputWithContext(ctx context.
 // DnsSearchPathsMapInput is an input type that accepts DnsSearchPathsMap and DnsSearchPathsMapOutput values.
 // You can construct a concrete instance of `DnsSearchPathsMapInput` via:
 //
-//          DnsSearchPathsMap{ "key": DnsSearchPathsArgs{...} }
+//	DnsSearchPathsMap{ "key": DnsSearchPathsArgs{...} }
 type DnsSearchPathsMapInput interface {
 	pulumi.Input
 
@@ -185,6 +188,11 @@ func (o DnsSearchPathsOutput) ToDnsSearchPathsOutput() DnsSearchPathsOutput {
 
 func (o DnsSearchPathsOutput) ToDnsSearchPathsOutputWithContext(ctx context.Context) DnsSearchPathsOutput {
 	return o
+}
+
+// Devices on your network will use these domain suffixes to resolve DNS names.
+func (o DnsSearchPathsOutput) SearchPaths() pulumi.StringArrayOutput {
+	return o.ApplyT(func(v *DnsSearchPaths) pulumi.StringArrayOutput { return v.SearchPaths }).(pulumi.StringArrayOutput)
 }
 
 type DnsSearchPathsArrayOutput struct{ *pulumi.OutputState }
