@@ -1,3 +1,3 @@
 // A Pulumi package for creating and managing xyz cloud resources.
-//
+
 package tailscale
