@@ -2,6 +2,7 @@
 # *** WARNING: this file was generated by the Pulumi Terraform Bridge (tfgen) Tool. ***
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
+import copy
 import warnings
 import pulumi
 import pulumi.runtime
@@ -43,6 +44,9 @@ class GetDeviceResult:
     @property
     @pulumi.getter
     def addresses(self) -> Sequence[str]:
+        """
+        The list of device's IPs
+        """
         return pulumi.get(self, "addresses")
 
     @property
@@ -56,21 +60,33 @@ class GetDeviceResult:
     @property
     @pulumi.getter
     def name(self) -> str:
+        """
+        The name of the device
+        """
         return pulumi.get(self, "name")
 
     @property
     @pulumi.getter
     def tags(self) -> Sequence[str]:
+        """
+        The tags applied to the device
+        """
         return pulumi.get(self, "tags")
 
     @property
     @pulumi.getter
     def user(self) -> str:
+        """
+        The user associated with the device
+        """
         return pulumi.get(self, "user")
 
     @property
     @pulumi.getter(name="waitFor")
     def wait_for(self) -> Optional[str]:
+        """
+        If specified, the provider will make multiple attempts to obtain the data source until the wait_for duration is reached. Retries are made every second so this value should be greater than 1s
+        """
         return pulumi.get(self, "wait_for")
 
 
@@ -103,14 +119,15 @@ def get_device(name: Optional[str] = None,
     sample_device = tailscale.get_device(name="user1-device.example.com",
         wait_for="60s")
     ```
+
+
+    :param str name: The name of the device
+    :param str wait_for: If specified, the provider will make multiple attempts to obtain the data source until the wait_for duration is reached. Retries are made every second so this value should be greater than 1s
     """
     __args__ = dict()
     __args__['name'] = name
     __args__['waitFor'] = wait_for
-    if opts is None:
-        opts = pulumi.InvokeOptions()
-    if opts.version is None:
-        opts.version = _utilities.get_version()
+    opts = pulumi.InvokeOptions.merge(_utilities.get_invoke_opts_defaults(), opts)
     __ret__ = pulumi.runtime.invoke('tailscale:index/getDevice:getDevice', __args__, opts=opts, typ=GetDeviceResult).value
 
     return AwaitableGetDeviceResult(
@@ -138,5 +155,9 @@ def get_device_output(name: Optional[pulumi.Input[str]] = None,
     sample_device = tailscale.get_device(name="user1-device.example.com",
         wait_for="60s")
     ```
+
+
+    :param str name: The name of the device
+    :param str wait_for: If specified, the provider will make multiple attempts to obtain the data source until the wait_for duration is reached. Retries are made every second so this value should be greater than 1s
     """
     ...
