@@ -2,6 +2,7 @@
 # *** WARNING: this file was generated by the Pulumi Terraform Bridge (tfgen) Tool. ***
 # *** Do not edit by hand unless you're certain you know what you are doing! ***
 
+import copy
 import warnings
 import pulumi
 import pulumi.runtime
@@ -35,6 +36,9 @@ class GetDevicesResult:
     @property
     @pulumi.getter
     def devices(self) -> Sequence['outputs.GetDevicesDeviceResult']:
+        """
+        The list of devices in the tailnet
+        """
         return pulumi.get(self, "devices")
 
     @property
@@ -48,6 +52,9 @@ class GetDevicesResult:
     @property
     @pulumi.getter(name="namePrefix")
     def name_prefix(self) -> Optional[str]:
+        """
+        Filters the device list to elements whose name has the provided prefix
+        """
         return pulumi.get(self, "name_prefix")
 
 
@@ -75,13 +82,13 @@ def get_devices(name_prefix: Optional[str] = None,
 
     sample_devices = tailscale.get_devices(name_prefix="example-")
     ```
+
+
+    :param str name_prefix: Filters the device list to elements whose name has the provided prefix
     """
     __args__ = dict()
     __args__['namePrefix'] = name_prefix
-    if opts is None:
-        opts = pulumi.InvokeOptions()
-    if opts.version is None:
-        opts.version = _utilities.get_version()
+    opts = pulumi.InvokeOptions.merge(_utilities.get_invoke_opts_defaults(), opts)
     __ret__ = pulumi.runtime.invoke('tailscale:index/getDevices:getDevices', __args__, opts=opts, typ=GetDevicesResult).value
 
     return AwaitableGetDevicesResult(
@@ -104,5 +111,8 @@ def get_devices_output(name_prefix: Optional[pulumi.Input[Optional[str]]] = None
 
     sample_devices = tailscale.get_devices(name_prefix="example-")
     ```
+
+
+    :param str name_prefix: Filters the device list to elements whose name has the provided prefix
     """
     ...
